+++ js/src/modules/anchors.js
@@ -3,27 +3,21 @@
  * @Author: Roni Laukkarinen
  * @Date:   2022-05-07 12:20:13
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-05-12 17:32:43
+ * @Last Modified time: 2022-09-28 14:33:09
  */
 import MoveTo from 'moveto';
 
 const initAnchors = () => {
   // General anchors used in links with class "js-trigger"
   const easeFunctions = {
-    easeInQuad(t, b, c, d) {
-      t /= d;
-      return c * t * t + b;
-    },
+    easeInQuad(t, b, c, d) { t /= d; return c * t * t + b; },
     easeOutQuad(t, b, c, d) {
-      t /= d;
-      return -c * t * (t - 2) + b;
+      t /= d; return -c * t * (t - 2) + b;
     },
   };
 
   const moveTo = new MoveTo(
-    {
-      ease: 'easeInQuad',
-    },
+    { ease: 'easeInQuad' },
     easeFunctions,
   );