+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add external-link-icon styles for mobile
 * Remove setup_editor_styles function because it breaks stage+prod and block styles still load correctly without that
+* Improve anchors.js formatting
 
 ### 9.1.6: 2022-09-14